+++ SpatialGDK/Source/SpatialGDKTests/SpatialGDK/EngineClasses/SpatialVirtualWorkerTranslationManager/SpatialVirtualWorkerTranslationManagerTest.cpp
@@ -55,7 +55,7 @@ VIRTUALWORKERTRANSLATIONMANAGER_TEST(Given_an_authority_change_THEN_query_for_wo
 	QueryOp.authority = WORKER_AUTHORITY_AUTHORITATIVE;
 
 	Manager->AuthorityChanged(QueryOp);
-	TestTrue("On gaining authority, the TranslationManager queried for worker entities.", Connection->GetLastEntityQuery() != nullptr);
+	TestTrue("On gaining authority, the TranslationManager queried for server worker entities.", Connection->GetLastEntityQuery() != nullptr);
 
 	EntityQueryDelegate* Delegate = Dispatcher->GetEntityQueryDelegate(0);
 	TestTrue("An EntityQueryDelegate was added to the dispatcher when the query was made", Delegate != nullptr);
@@ -86,7 +86,7 @@ VIRTUALWORKERTRANSLATIONMANAGER_TEST(Given_a_failed_query_response_THEN_query_ag
 	Manager->AddVirtualWorkerIds(VirtualWorkerIds);
 
 	Delegate->ExecuteIfBound(ResponseOp);
-	TestTrue("After a failed query response, the TranslationManager queried again for worker entities.", Connection->GetLastEntityQuery() != nullptr);
+	TestTrue("After a failed query response, the TranslationManager queried again for server worker entities.", Connection->GetLastEntityQuery() != nullptr);
 
 	return true;
 }
@@ -111,7 +111,7 @@ VIRTUALWORKERTRANSLATIONMANAGER_TEST(Given_a_successful_query_without_enough_wor
 	Manager->AddVirtualWorkerIds(VirtualWorkerIds);
 
 	Delegate->ExecuteIfBound(ResponseOp);
-	TestTrue("When not enough workers available, the TranslationManager queried again for worker entities.", Connection->GetLastEntityQuery() != nullptr);
+	TestTrue("When not enough workers available, the TranslationManager queried again for server worker entities.", Connection->GetLastEntityQuery() != nullptr);
 
 	return true;
 }
@@ -142,7 +142,7 @@ VIRTUALWORKERTRANSLATIONMANAGER_TEST(Given_a_successful_query_with_invalid_worke
 	Manager->AddVirtualWorkerIds(VirtualWorkerIds);
 
 	Delegate->ExecuteIfBound(ResponseOp);
-	TestTrue("When enough workers available but they are invalid, the TranslationManager queried again for worker entities.", Connection->GetLastEntityQuery() != nullptr);
+	TestTrue("When enough workers available but they are invalid, the TranslationManager queried again for server worker entities.", Connection->GetLastEntityQuery() != nullptr);
 
 	return true;
 }